requirements/common.txt
Django>=1.11
psycopg2==2.7
pytz==2017.3
django_cors_headers==2
djangorestframework==3.7
whitenoise==3.1
Pillow==4